changes from PR
