+++ pkg/graveler/committed/merge.go
@@ -18,11 +18,10 @@ type merger struct {
 	source               Iterator
 	dest                 Iterator
 	haveSource, haveDest bool
-	summary              graveler.DiffSummary
 }
 
-// baseGE returns range from base iterator, which is greater or equal than the given key
-func (m *merger) baseRangeGE(key graveler.Key) (*Range, error) {
+// moveBaseToGERange moves base iterator (from current point) to range which is greater or equal than the given key
+func (m *merger) moveBaseToGERange(key graveler.Key) (*Range, error) {
 	for {
 		_, baseRange := m.base.Value()
 		if baseRange != nil && bytes.Compare(baseRange.MaxKey, key) >= 0 {
@@ -32,39 +31,29 @@ func (m *merger) baseRangeGE(key graveler.Key) (*Range, error) {
 			break
 		}
 	}
-	if err := m.base.Err(); err != nil {
-		return nil, err
-	}
-	return nil, nil
+	return nil, m.base.Err()
 }
 
-// baseGE returns value from base iterator, which is greater or equal than the given key
-func (m *merger) baseKeyGE(key graveler.Key) (*graveler.ValueRecord, error) {
+// moveBaseToGEKey moves base iterator (from current point) to key from is greater or equal than the given key
+func (m *merger) moveBaseToGEKey(key graveler.Key) (*graveler.ValueRecord, error) {
 	baseValue, _ := m.base.Value()
 	if baseValue != nil && bytes.Compare(key, baseValue.Key) <= 0 {
 		return baseValue, nil
 	}
-
+	baseRange, err := m.moveBaseToGERange(key)
+	if err != nil {
+		return nil, err
+	}
 	for {
-		_, baseRange := m.base.Value()
-		if baseRange != nil && bytes.Compare(baseRange.MaxKey, key) >= 0 {
-			for {
-				baseValue, innerRange := m.base.Value()
-				if baseValue != nil && bytes.Compare(key, baseValue.Key) <= 0 {
-					return baseValue, nil
-				}
-				if !m.base.Next() || innerRange.ID != baseRange.ID {
-					break
-				}
-			}
-		} else if !m.base.NextRange() {
+		baseValue, innerRange := m.base.Value()
+		if baseValue != nil && bytes.Compare(key, baseValue.Key) <= 0 {
+			return baseValue, nil
+		}
+		if !m.base.Next() || innerRange.ID != baseRange.ID {
 			break
 		}
 	}
-	if err := m.base.Err(); err != nil {
-		return nil, err
-	}
-	return nil, nil
+	return nil, m.base.Err()
 }
 
 // writeRange writes Range using writer
@@ -96,8 +85,8 @@ func (m *merger) writeRecord(writeValue *graveler.ValueRecord) error {
 	return nil
 }
 
-// applyAll applies all remaining changes from Iterator to writer
-func (m *merger) applyAll(iter Iterator) error {
+// handleAll handles the case where only one Iterator from source or dest remains
+func (m *merger) handleAll(iter Iterator) error {
 	for {
 		select {
 		case <-m.ctx.Done():
@@ -106,7 +95,7 @@ func (m *merger) applyAll(iter Iterator) error {
 		}
 		iterValue, iterRange := iter.Value()
 		if iterValue == nil {
-			baseRange, err := m.baseRangeGE(graveler.Key(iterRange.MinKey))
+			baseRange, err := m.moveBaseToGERange(graveler.Key(iterRange.MinKey))
 			if err != nil {
 				return fmt.Errorf("base range GE: %w", err)
 			}
@@ -119,7 +108,7 @@ func (m *merger) applyAll(iter Iterator) error {
 				break
 			}
 		} else {
-			baseValue, err := m.baseKeyGE(iterValue.Key)
+			baseValue, err := m.moveBaseToGEKey(iterValue.Key)
 			if err != nil {
 				return fmt.Errorf("base value GE: %w", err)
 			}
@@ -136,10 +125,10 @@ func (m *merger) applyAll(iter Iterator) error {
 	return iter.Err()
 }
 
-// applyBothRanges applies merging when both source and dest on range header
-func (m *merger) applyBothRanges(sourceRange *Range, destRange *Range) error {
+// handleBothRanges handles the case where both source and dest iterators are at the header of a range
+func (m *merger) handleBothRanges(sourceRange *Range, destRange *Range) error {
 	switch {
-	case sourceRange.ID == destRange.ID: // source and dest added the same range
+	case sourceRange.ID == destRange.ID: // range hasn't changed or both added the same range
 		err := m.writeRange(sourceRange)
 		if err != nil {
 			return err
@@ -148,7 +137,7 @@ func (m *merger) applyBothRanges(sourceRange *Range, destRange *Range) error {
 		m.haveDest = m.dest.NextRange()
 
 	case bytes.Compare(sourceRange.MaxKey, destRange.MinKey) < 0: // source before dest
-		baseRange, err := m.baseRangeGE(graveler.Key(sourceRange.MinKey))
+		baseRange, err := m.moveBaseToGERange(graveler.Key(sourceRange.MinKey))
 		if err != nil {
 			return fmt.Errorf("base range GE: %w", err)
 		}
@@ -164,10 +153,12 @@ func (m *merger) applyBothRanges(sourceRange *Range, destRange *Range) error {
 			m.haveSource = m.source.NextRange()
 			return nil
 		}
-		return graveler.ErrConflictFound // conflict - both changed this range
+		// both changed this range
+		m.haveSource = m.source.Next()
+		m.haveDest = m.dest.Next()
 
 	case bytes.Compare(destRange.MaxKey, sourceRange.MinKey) < 0: // dest before source
-		baseRange, err := m.baseRangeGE(graveler.Key(destRange.MinKey))
+		baseRange, err := m.moveBaseToGERange(graveler.Key(destRange.MinKey))
 		if err != nil {
 			return fmt.Errorf("base range GE: %w", err)
 		}
@@ -183,10 +174,12 @@ func (m *merger) applyBothRanges(sourceRange *Range, destRange *Range) error {
 			m.haveDest = m.dest.NextRange()
 			return nil
 		}
-		return graveler.ErrConflictFound // conflict - both changed this range
+		// both changed this range
+		m.haveSource = m.source.Next()
+		m.haveDest = m.dest.Next()
 
 	case bytes.Equal(sourceRange.MinKey, destRange.MinKey) && bytes.Equal(sourceRange.MaxKey, destRange.MaxKey): // same bounds
-		baseRange, err := m.baseRangeGE(graveler.Key(sourceRange.MinKey))
+		baseRange, err := m.moveBaseToGERange(graveler.Key(sourceRange.MinKey))
 		if err != nil {
 			return err
 		}
@@ -213,19 +206,19 @@ func (m *merger) applyBothRanges(sourceRange *Range, destRange *Range) error {
 	return nil
 }
 
-// applyBothKeys applies merging when both source and dest insides ranges
-func (m *merger) applyBothKeys(sourceValue *graveler.ValueRecord, destValue *graveler.ValueRecord) error {
+// handleBothKeys handles the case where both source and dest iterators are inside range
+func (m *merger) handleBothKeys(sourceValue *graveler.ValueRecord, destValue *graveler.ValueRecord) error {
 	c := bytes.Compare(sourceValue.Key, destValue.Key)
 	switch {
 	case c < 0: // source before dest
-		baseValue, err := m.baseKeyGE(sourceValue.Key)
+		baseValue, err := m.moveBaseToGEKey(sourceValue.Key)
 		if err != nil {
 			return err
 		}
 		if baseValue != nil && bytes.Equal(sourceValue.Identity, baseValue.Identity) { // dest deleted this record
 			m.haveSource = m.source.Next()
 		} else {
-			if baseValue != nil && bytes.Equal(sourceValue.Key, baseValue.Key) { // conflict
+			if baseValue != nil && bytes.Equal(sourceValue.Key, baseValue.Key) { // deleted by dest and changed by source
 				return graveler.ErrConflictFound
 			}
 			// source added this record
@@ -236,14 +229,14 @@ func (m *merger) applyBothKeys(sourceValue *graveler.ValueRecord, destValue *gra
 			m.haveSource = m.source.Next()
 		}
 	case c > 0: // dest before source
-		baseValue, err := m.baseKeyGE(destValue.Key)
+		baseValue, err := m.moveBaseToGEKey(destValue.Key)
 		if err != nil {
 			return err
 		}
 		if baseValue != nil && bytes.Equal(destValue.Identity, baseValue.Identity) { // source deleted this record
 			m.haveDest = m.dest.Next()
 		} else {
-			if baseValue != nil && bytes.Equal(destValue.Key, baseValue.Key) { // conflict
+			if baseValue != nil && bytes.Equal(destValue.Key, baseValue.Key) { // deleted by source added by dest
 				return graveler.ErrConflictFound
 			}
 			// dest added this record
@@ -254,25 +247,43 @@ func (m *merger) applyBothKeys(sourceValue *graveler.ValueRecord, destValue *gra
 			m.haveDest = m.dest.Next()
 		}
 	default: // identical keys
-		baseValue, err := m.baseKeyGE(destValue.Key)
+		baseValue, err := m.moveBaseToGEKey(destValue.Key)
 		if err != nil {
 			return err
 		}
 		if !bytes.Equal(sourceValue.Identity, destValue.Identity) {
-			if baseValue != nil && (bytes.Equal(sourceValue.Identity, baseValue.Identity) ||
-				bytes.Equal(destValue.Identity, baseValue.Identity)) {
-				err := m.writeRecord(sourceValue)
+			if baseValue != nil {
+				if bytes.Equal(sourceValue.Identity, baseValue.Identity) {
+					err = m.writeRecord(destValue)
+				} else if bytes.Equal(destValue.Identity, baseValue.Identity) {
+					err = m.writeRecord(sourceValue)
+				} else { // both changed the same key
+					return graveler.ErrConflictFound
+				}
 				if err != nil {
 					return fmt.Errorf("write record: %w", err)
 				}
 				m.haveSource = m.source.Next()
 				m.haveDest = m.dest.Next()
 				return nil
-			} else { // both changed the same key
+			} else {
 				return graveler.ErrConflictFound
 			}
+
+			//if baseValue != nil && (bytes.Equal(sourceValue.Identity, baseValue.Identity) ||
+			//	bytes.Equal(destValue.Identity, baseValue.Identity)) {
+			//	err := m.writeRecord(sourceValue)
+			//	if err != nil {
+			//		return fmt.Errorf("write record: %w", err)
+			//	}
+			//	m.haveSource = m.source.Next()
+			//	m.haveDest = m.dest.Next()
+			//	return nil
+			//} else { // both changed the same key
+			//	return graveler.ErrConflictFound
+			//}
 		}
-		// both added the same record
+		// record hasn't changed or both added the same record
 		err = m.writeRecord(sourceValue)
 		if err != nil {
 			return fmt.Errorf("write record: %w", err)
@@ -283,17 +294,17 @@ func (m *merger) applyBothKeys(sourceValue *graveler.ValueRecord, destValue *gra
 	return nil
 }
 
-// applyDestRangeSourceKey applies merging when source inside range and dest on range header
-func (m *merger) applyDestRangeSourceKey(destRange *Range, sourceValue *graveler.ValueRecord) error {
+// handleDestRangeSourceKey handles the case where source Iterator inside range and dest Iterator at the header of a range
+func (m *merger) handleDestRangeSourceKey(destRange *Range, sourceValue *graveler.ValueRecord) error {
 	if bytes.Compare(destRange.MinKey, sourceValue.Key) > 0 { // source before dest range
-		baseValue, err := m.baseKeyGE(sourceValue.Key)
+		baseValue, err := m.moveBaseToGEKey(sourceValue.Key)
 		if err != nil {
 			return fmt.Errorf("base key GE: %w", err)
 		}
 		if baseValue != nil && bytes.Equal(sourceValue.Identity, baseValue.Identity) { // dest deleted this record
 			m.haveSource = m.source.Next()
 		} else {
-			if baseValue != nil && bytes.Equal(sourceValue.Key, baseValue.Key) { // conflict
+			if baseValue != nil && bytes.Equal(sourceValue.Key, baseValue.Key) { // deleted by dest and changed by source
 				return graveler.ErrConflictFound
 			}
 			// source added this record
@@ -307,15 +318,13 @@ func (m *merger) applyDestRangeSourceKey(destRange *Range, sourceValue *graveler
 	}
 
 	if bytes.Compare(destRange.MaxKey, sourceValue.Key) < 0 { // dest range before source
-		baseRange, err := m.baseRangeGE(graveler.Key(destRange.MinKey))
+		baseRange, err := m.moveBaseToGERange(graveler.Key(destRange.MinKey))
 		if err != nil {
 			return fmt.Errorf("base range GE: %w", err)
 		}
 		if baseRange != nil && destRange.ID == baseRange.ID { // source deleted this range
 			m.haveDest = m.dest.NextRange()
 			return nil
-		} else {
-			return graveler.ErrConflictFound // conflict
 		}
 	}
 	// dest is at start of range which we need to scan, enter it
@@ -323,17 +332,17 @@ func (m *merger) applyDestRangeSourceKey(destRange *Range, sourceValue *graveler
 	return nil
 }
 
-// applySourceRangeDestKey applies merging when dest inside range and source on range header
-func (m *merger) applySourceRangeDestKey(sourceRange *Range, destValue *graveler.ValueRecord) error {
+// handleSourceRangeDestKey handles the case where dest Iterator inside range and source Iterator at the header of a range
+func (m *merger) handleSourceRangeDestKey(sourceRange *Range, destValue *graveler.ValueRecord) error {
 	if bytes.Compare(sourceRange.MinKey, destValue.Key) > 0 { // dest before source range
-		baseValue, err := m.baseKeyGE(destValue.Key)
+		baseValue, err := m.moveBaseToGEKey(destValue.Key)
 		if err != nil {
 			return fmt.Errorf("base key GE: %w", err)
 		}
 		if baseValue != nil && bytes.Equal(destValue.Identity, baseValue.Identity) { // source deleted this record
 			m.haveSource = m.source.Next()
 		} else {
-			if baseValue != nil && bytes.Equal(destValue.Key, baseValue.Key) { // conflict
+			if baseValue != nil && bytes.Equal(destValue.Key, baseValue.Key) { // deleted by source and changed by dest
 				return graveler.ErrConflictFound
 			}
 			// dest added this record
@@ -347,15 +356,13 @@ func (m *merger) applySourceRangeDestKey(sourceRange *Range, destValue *graveler
 	}
 
 	if bytes.Compare(sourceRange.MaxKey, destValue.Key) < 0 { // source range before dest
-		baseRange, err := m.baseRangeGE(graveler.Key(sourceRange.MinKey))
+		baseRange, err := m.moveBaseToGERange(graveler.Key(sourceRange.MinKey))
 		if err != nil {
 			return fmt.Errorf("base range GE: %w", err)
 		}
 		if baseRange != nil && sourceRange.ID == baseRange.ID { // dest deleted this range
 			m.haveSource = m.source.NextRange()
 			return nil
-		} else {
-			return graveler.ErrConflictFound // conflict
 		}
 	}
 	// source is at start of range which we need to scan, enter it
@@ -376,13 +383,13 @@ func (m *merger) merge() error {
 		var err error
 		switch {
 		case sourceValue == nil && destValue == nil:
-			err = m.applyBothRanges(sourceRange, destRange)
+			err = m.handleBothRanges(sourceRange, destRange)
 		case destValue == nil && sourceValue != nil:
-			err = m.applyDestRangeSourceKey(destRange, sourceValue)
+			err = m.handleDestRangeSourceKey(destRange, sourceValue)
 		case sourceValue == nil && destValue != nil:
-			err = m.applySourceRangeDestKey(sourceRange, destValue)
+			err = m.handleSourceRangeDestKey(sourceRange, destValue)
 		default:
-			err = m.applyBothKeys(sourceValue, destValue)
+			err = m.handleBothKeys(sourceValue, destValue)
 		}
 		if err != nil {
 			return err
@@ -399,27 +406,26 @@ func (m *merger) merge() error {
 	}
 
 	if m.haveSource {
-		if err := m.applyAll(m.source); err != nil {
+		if err := m.handleAll(m.source); err != nil {
 			return err
 		}
 	}
 	if m.haveDest {
-		if err := m.applyAll(m.dest); err != nil {
+		if err := m.handleAll(m.dest); err != nil {
 			return err
 		}
 	}
 	return nil
 }
 
-func Merge(ctx context.Context, writer MetaRangeWriter, base Iterator, source Iterator, destination Iterator) (graveler.DiffSummary, error) {
+func Merge(ctx context.Context, writer MetaRangeWriter, base Iterator, source Iterator, destination Iterator) error {
 	m := merger{
-		ctx:     ctx,
-		logger:  logging.FromContext(ctx),
-		writer:  writer,
-		base:    base,
-		source:  source,
-		dest:    destination,
-		summary: graveler.DiffSummary{Count: make(map[graveler.DiffType]int), Incomplete: true}, // Incomplete because skipping ranges
+		ctx:    ctx,
+		logger: logging.FromContext(ctx),
+		writer: writer,
+		base:   base,
+		source: source,
+		dest:   destination,
 	}
-	return m.summary, m.merge()
+	return m.merge()
 }
